[ISSUE #9630]♻️Partition RemotingCommand frame encoding and decoding
